Fix ifstream/ofstream ambiguity bug
Many files include both `fstream` and `boost/filesystem.hpp`.
This becomes a problem when we use both namespaces `std` and
`boost::filesystem` because both define `ifstream` and `ofstream`.
This caused a compile-time error:
`error: reference to 'ifstream' is ambiguous`

This appends `std::` to `ifstream`/`ofstream` when necessary.
(This is consistent with the rest of the project, which uses
the `std` version and not the `boost::filesystem` version.)

(Bug noticed with gcc v5.3.0 and boost v1.60.0 on CentOS v6.2)