don't build slug lib by default
Typing `make` (in any directory) will do the same as it did before
there was an option to build a shared library.

In order to get a shared library for `slug` you now need to go into
the `src` directory and use `make lib` or `make all` (or `make debuglib`
if you need a debug-friendly library).

(This might be a little confusing; using `make all` in the root
directory won't actually make all targets)